+++ deny.toml
@@ -79,6 +79,7 @@ allow = [
     "BSD-3-Clause",
     "CC0-1.0",
     "BSD-2-Clause",
+    "MPL-2.0",
 ]
 # List of explicitly disallowed licenses
 # See https://spdx.org/licenses/ for list of possible licenses
