+++ examples/candle_whisper/Cargo.toml
@@ -10,13 +10,26 @@ version.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
-metal = ["candle-core/metal"]
+metal = ["candle-core/metal", "candle-nn/metal"]
+cuda = ["candle-core/cuda", "candle-nn/cuda", "candle-transformers/cuda", "dep:bindgen_cuda"]
+
+
+[build-dependencies]
+anyhow = { version = "1", features = ["backtrace"] }
+bindgen_cuda = { version = "0.1.1", optional = true }
+
 
 [dependencies]
+clap = { version = "4.2.4", features = ["derive"] }
 serde = { version = "1.0", features = ["derive"] }
 silent = { path = "../../silent", features = ["full"] }
 symphonia = { version = "0.5.3", features = ["all"] }
 anyhow = "1.0.79"
+
 candle-core = { version = "0.3.2", optional = true }
 candle-nn = { version = "0.3.2", optional = true }
-
+candle-transformers = { version = "0.3.2", optional = true }
+tokenizers = { version = "0.13.4", default-features = false }
+rand = "0.8.5"
+serde_json = "1.0.109"
+byteorder = "1.5.0"
